move the gitbook package under commands/docs/
