+++ commands/docs/docs.go
@@ -4,7 +4,7 @@ import (
 	"os"
 
 	"github.com/opencontrol/compliance-masonry/docx"
-	"github.com/opencontrol/compliance-masonry/gitbook"
+	"github.com/opencontrol/compliance-masonry/commands/docs/gitbook"
 	"github.com/opencontrol/compliance-masonry/tools/certifications"
 	"errors"
 )
